+++ hydromt_wflow/wflow.py
@@ -3911,8 +3911,7 @@ Run setup_soilmaps first"
         connection_method: str = "subbasin_area",
         area_max: float = 30.0,
         basin_buffer_cells: int = 0,
-        geom_snapping_tolerance: float = 0.0,
-        min_stream_order: int = 2,
+        geom_snapping_tolerance: float = 0.001,
         add_tributaries: bool = True,
         include_river_boundaries: bool = True,
         mapname: str = "1dmodel",
@@ -3947,6 +3946,10 @@ Run setup_soilmaps first"
         save all river inflows for the subcatchments and lateral.river.q_av for the
         tributaries using :py:meth:`hydromt_wflow.wflow.setup_config_output_timeseries`.
 
+        Note that if the method is `subbasin_area`, only connecting one continuous
+        river1d is supported. To connect several riv1d networks to the same wflow model,
+        the method can be run several times.
+
         Adds model layer:
 
         * **subcatchment_{mapname}** map/geom:  connection subbasins between
@@ -3969,14 +3972,13 @@ Run setup_soilmaps first"
             connection_method **subbasin_area** or **nodes** with add_tributaries
             set to True.
         basin_buffer_cells : int, default 0
-            Number of cells to use when clipping the river geometry to the basin extent.
-            This can be used to not include river geometries near the basin border.
-        geom_snapping_tolerance : float, default 0.0
-            Distance used to determine whether to snap parts of the river geometry that
-            are close to each other.
-        min_stream_order: int, default 2
-            Minimum stream order of the river cells to connect to the 1D model. Includes
-            all river cells when set to 1.
+            Number of cells to use when clipping the river1d geometry to the basin
+            extent. This can be used to not include river geometries near the basin
+            border.
+        geom_snapping_tolerance : float, default 0.1
+            Distance used to determine whether to snap parts of the river1d geometry
+            that are close to each other. This can be useful if some of the tributaries
+            of the 1D river are not perfectly connected to the main river.
         add_tributaries : bool, default True
             If True, derive tributaries for the subbasins larger than area_max. Always
             True for **subbasin_area** method.
@@ -4000,6 +4002,7 @@ Run setup_soilmaps first"
         See Also
         --------
         hydromt.flw.gauge_map
+        hydromt_wflow.workflows.wflow_1dmodel_connection
         """
         # Check connection method values
         if connection_method not in ["subbasin_area", "nodes"]:
@@ -4023,7 +4026,6 @@ Run setup_soilmaps first"
             area_max=area_max,
             basin_buffer_cells=basin_buffer_cells,
             geom_snapping_tolerance=geom_snapping_tolerance,
-            min_stream_order=min_stream_order,
             add_tributaries=add_tributaries,
             include_river_boundaries=include_river_boundaries,
             logger=self.logger,